Split out limits to new file from smart playlist functionality
